requirements.txt
pyyaml
numpy
matplotlib
opencv-python
scikit-learn
pyqt6